review small fixes
- changed match to endsWith
- getImportStatements return a typed ReadonlyArray instead of clone
- inline context object in transformer
- rename symbol name for st-import feature
- reorder st-import feature diagnostics (spread first)
- add keyframes import test for legacy pseudo import
- move 2 value functions from pseudo-states to helpers/value
